+++ CHANGELOG.md
@@ -3,5 +3,9 @@ Insanely configurable logging for Node.js
 
 ##ChangeLog
 
+### Development HEAD
+- Fix: Remove comma separator from date elements in JSON formatter
+- Added: Human formatter
+
 ### v0.1.0
 - **Initial Release**